remove last use of adblock-rust's list API
