+++ lib/adBlockRustUtils.js
@@ -11,6 +11,32 @@ const uBlockScriptlets = path.join(uBlockLocalRoot, 'assets/resources/scriptlets
 
 const braveResourcesUrl = 'https://raw.githubusercontent.com/brave/adblock-resources/master/dist/resources.json'
 
+const defaultListsUrl = 'https://raw.githubusercontent.com/brave/adblock-resources/master/filter_lists/default.json'
+const regionalListsUrl = 'https://raw.githubusercontent.com/brave/adblock-resources/master/filter_lists/regional.json'
+
+/**
+ * Returns a promise that which resolves with the body parsed as JSON
+ *
+ * @param url The URL to fetch from
+ * @return a promise that resolves with the content of the list or rejects with an error message.
+ */
+const requestJSON = (url) => new Promise((resolve, reject) => {
+  request.get(url, function (error, response, body) {
+    if (error) {
+      reject(new Error(`Request error: ${error}`))
+      return
+    }
+    if (response.statusCode !== 200) {
+      reject(new Error(`Error status code ${response.statusCode} returned for URL: ${url}`))
+      return
+    }
+    resolve(JSON.parse(body))
+  })
+})
+
+const getDefaultLists = requestJSON.bind(null, defaultListsUrl)
+const getRegionalLists = requestJSON.bind(null, regionalListsUrl)
+
 /**
  * Returns a promise that generates a resources file from the uBlock Origin
  * repo hosted on GitHub
@@ -38,3 +64,5 @@ const generateResourcesFile = (outLocation) => {
 }
 
 module.exports.generateResourcesFile = generateResourcesFile
+module.exports.getDefaultLists = getDefaultLists
+module.exports.getRegionalLists = getRegionalLists
